requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-modeltranslation==0.18.9
djangorestframework==3.14.0
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
Pillow==9.5.0
psycopg2==2.9.6
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
sqlparse==0.4.4
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==2.0.2
